requirements.txt
Django==4.1.7
django-allauth==0.51.0
django-filter==22.1
django-js-asset==2.0.0
django-ratelimit==4.0.0
django-redis==5.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
email-validator==1.3.1
Pillow==9.2.0
psycopg2==2.9.4
python-dotenv==0.21.0
python-multipart==0.0.5
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==2.0.3
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
psycopg2-binary==2.9.5
python-decouple==3.8
gunicorn==20.1.0
django-simple-captcha==0.5.17